requirements.txt
# Use GPU version of torch for a faster performance, it will depend on your system configuration
# For example, for CUDA 11.7, use the following line:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu117
# By default, it will install the CPU version
torch
torchvision
open-clip-torch 
faiss-cpu
numpy
Pillow
tqdm
opencv-python
matplotlib
ultralytics
google-genai
python-dotenv
